If we can't write yet, don't bother pulling off a packet
This logic previously could cause the fdcanusb to "drop" packets, if
one came in while it was in the middle of sending something to the USB
bus.  Now it will at least use the FDCAN peripheral's buffers to hold
the frames until the USB write is complete.  Overruns are still
possible, but at least our buffer size is non-zero.