Basic support for also enumerating as a gs_usb device
This means that the fdcanusb/mjcanfd-usb-1x will present as a
"composite USB" device with two distinct IADs.  One emulates the
gs_usb protocol used by the Linux gs_usb driver, the other is a CDC
interface as before.

The two interfaces are designed to be usable in parallel or
separately, although if used in parallel, some things like bus/timing
configuration will stomp on each either.

Sufficient USB descriptors are present to enable the device to be used
on Windows without a dedicated driver or .inf file, although the
python-can gs_usb interface is lacking and does not support many
features like CAN-FD or sample point configuration among others.

To use with the Linux gs_usb driver, an appropriate `ip` command can
be used, like the following for moteus-compatible timings:

```
ip link set can0 up type can bitrate 1000000 dbitrate 5000000 fd on \
  sjw 10 dsjw 5 sample-point 0.666 dsample-point 0.666 \
  restart-ms 1000
```

Not all features of the gs_usb protocol are implemented yet.